@@ -4,86 +4,86 @@
   "private": true,
   "dependencies": {
     "@date-io/moment": "^1.3.13",
-    "@material-ui/core": "^4.11.4",
+    "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/pickers": "^3.2.10",
-    "@sentry/browser": "^6.7.1",
-    "@types/js-cookie": "^2.2.6",
+    "@material-ui/pickers": "^3.3.10",
+    "@sentry/browser": "^6.13.3",
+    "@types/js-cookie": "^2.2.7",
     "bluebird": "^3.7.2",
     "bootstrap-social": "^5.1.1",
     "connected-react-router": "^6.9.1",
     "export-to-csv": "^0.2.1",
-    "final-form": "^4.20.2",
+    "final-form": "^4.20.4",
     "final-form-material-ui": "^0.3.0",
-    "firebase": "^9.1.3",
+    "firebase": "^9.2.0",
     "fuzzy-search": "^3.2.1",
-    "google-libphonenumber": "^3.2.21",
+    "google-libphonenumber": "^3.2.25",
     "google-maps-react": "^2.0.6",
-    "gravatar": "^1.8.1",
+    "gravatar": "^1.8.2",
     "history": "^4.10.1",
-    "immutable": "^4.0.0-rc.12",
+    "immutable": "^4.0.0",
     "initials": "^3.1.1",
     "isemail": "^3.2.0",
     "jquery": "^3.6.0",
     "logrocket": "^1.3.0",
     "logrocket-react": "^4.0.1",
     "moment": "^2.29.1",
-    "mui-datatables": "^3.7.7",
-    "node-sass": "^4.14.1",
+    "mui-datatables": "^3.8.2",
     "normalize-email": "^1.1.1",
-    "parse-full-name": "^1.2.5",
+    "parse-full-name": "^1.2.6",
     "prop-types": "^15.7.2",
     "query-string": "^6.14.1",
     "react": "^16.14.0",
-    "react-bootstrap": "^1.6.1",
-    "react-copy-to-clipboard": "^5.0.3",
+    "react-bootstrap": "^1.6.4",
+    "react-copy-to-clipboard": "^5.0.4",
     "react-dfb": "^1.0.4",
     "react-dom": "^16.14.0",
-    "react-final-form": "^6.5.3",
-    "react-redux": "^7.2.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-final-form": "^6.5.7",
+    "react-redux": "^7.2.6",
+    "react-router": "^5.2.1",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "^4.0.3",
-    "react-scroll": "^1.8.2",
+    "react-scroll": "^1.8.4",
     "react-slick": "^0.28.1",
     "react-stripe-elements": "^6.1.2",
     "reactjs-file-uploader": "^1.0.8",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.0",
     "request-promise": "^4.2.6",
-    "typescript": "^4.2.4",
+    "sass": "^1.43.4",
+    "typescript": "^4.4.4",
     "underscore": "^1.13.1",
     "underscore.string": "^3.3.5",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@types/fuzzy-search": "^2.1.0",
+    "@types/fuzzy-search": "^2.1.1",
     "@types/google-libphonenumber": "^7.4.21",
     "@types/google-maps-react": "^2.0.5",
-    "@types/gravatar": "^1.8.2",
-    "@types/jest": "^26.0.23",
-    "@types/jquery": "^3.5.5",
-    "@types/mui-datatables": "^3.7.4",
-    "@types/node": "^14.14.31",
-    "@types/react": "^16.14.4",
+    "@types/gravatar": "^1.8.3",
+    "@types/jest": "^26.0.24",
+    "@types/jquery": "^3.5.8",
+    "@types/mui-datatables": "^3.7.7",
+    "@types/node": "^14.17.32",
+    "@types/react": "^16.14.20",
     "@types/react-copy-to-clipboard": "^4.3.0",
-    "@types/react-dom": "^16.9.11",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-scroll": "^1.8.2",
-    "@types/react-slick": "^0.23.4",
-    "@types/react-stripe-elements": "^6.0.4",
-    "@types/request-promise": "^4.1.47",
-    "@types/underscore": "^1.11.2",
+    "@types/react-dom": "^16.9.14",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/react-scroll": "^1.8.3",
+    "@types/react-slick": "^0.23.7",
+    "@types/react-stripe-elements": "^6.0.6",
+    "@types/request-promise": "^4.1.48",
+    "@types/underscore": "^1.11.3",
     "@types/underscore.string": "^0.0.38",
-    "@types/uuid": "^8.3.0",
+    "@types/uuid": "^8.3.1",
     "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-promise": "^5.1.1",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-standard": "^5.0.0",
-    "prettier": "^2.3.1",
+    "prettier": "^2.4.1",
     "tslint": "^6.1.3"
   },
   "scripts": {
