@@ -1,6 +1,5 @@
 import * as Admin from 'firebase-admin'
-
-const BPromise = require('bluebird')
+import { each } from 'bluebird'
 
 const GetPhotoUrl = (admin: Admin.app.App) => {
   const auth = admin.auth()
@@ -8,7 +7,7 @@ const GetPhotoUrl = (admin: Admin.app.App) => {
   const listAllUsers = async (nextPageToken?: string) => {
     // List batch of users, 1000 at a time.
     const listUsersResult = await auth.listUsers(1000, nextPageToken)
-    await BPromise.each(listUsersResult.users, async (userRecord: Admin.auth.UserRecord) => {
+    await each(listUsersResult.users, async (userRecord: Admin.auth.UserRecord) => {
       try {
         const foundProviderData = userRecord.providerData.find(({ photoURL }) => {
           return Boolean(photoURL)
